Separate the tests in service and action to avoid conflicts.
