+++ zenoh-test-ros2dds/Cargo.toml
@@ -22,7 +22,6 @@ r2r = "0.9"
 serde = "1.0.154"
 serde_derive = "1.0.154"
 serde_json = "1.0.114"
-serial_test = "3.2.0"
 tokio = { version = "1.35.1", default-features = false } # Default features are disabled due to some crates' requirements
 zenoh = { version = "1.0.0-dev", git = "https://github.com/eclipse-zenoh/zenoh.git", branch = "main", features = [
   "plugins",
